pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "filetype>=1.2.0",
    "puremagic>=1.30",
    "python-magic>=0.4.27",
]

[project.scripts]
file_val = "utils.file_val:cli"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
